Split domain model tests by entity type
- test_base.py: Base model and mixin tests
- test_sources.py: ImportSource/ImportRun tests
- test_people.py: Person/PersonAlias tests
- test_media.py: Media/MediaEmbedding tests
- test_messages.py: ChatThread/ChatMessage tests
- test_email.py: Email/EmailThread tests
- test_social.py: SocialPost/Subscription tests
- test_browsing.py: BrowsingHistory/Bookmark tests
- test_notes.py: KnowledgeNote tests
- test_locations.py: Location/LocationVisit tests
- test_calendar.py: CalendarEvent tests
- test_financial.py: Account/Transaction tests
- test_links.py: EntityLink tests
- test_tags.py: Tag/TagAssignment tests
